Use pip-tools, upgrade to Python 3.7

+++ Makefile
@@ -5,7 +5,8 @@ all: env node static
 .PHONY: env
 env:
 	[ -e server/env/bin/python ] || virtualenv --python=python3 server/env/
-	server/env/bin/pip install -q -r server/requirements.txt
+	[ -e server/env/bin/python ] || server/env/bin/pip install -q -r server/requirements.txt
+	server/env/bin/pip-sync
 
 .PHONY: node
 node:
